tests: Added tests for default encoding.
Also did some refactoring; moved some tests functionalities to utils so other test files can resort to it to use these functions.
