+++ tests/fixtures/log_trace_off.in.lsp
@@ -1,0 +1,12 @@
+Content-Length: 72
+
+{"jsonrpc":"2.0","id":1,"method":"initialize","params":{"trace":"off"}}
+Content-Length: 41
+
+{"jsonrpc":"2.0","method":"initialized"}
+Content-Length: 45
+
+{"jsonrpc":"2.0","id":2,"method":"shutdown"}
+Content-Length: 34
+
+{"jsonrpc":"2.0","method":"exit"}
