+++ tests/util.py
@@ -1,5 +1,7 @@
+import asyncio
 import io
 import json
+import typing
 
 from pylsp_jsonrpc.endpoint import Endpoint
 from pylsp_jsonrpc.exceptions import JsonRpcException
@@ -46,3 +48,44 @@ class FakeLanguageServer(MALLSPServer):
     def __init__(self, *args, **kwargs):
         kwargs.setdefault("EndpointClass", FakeEndpoint)
         super().__init__(*args, **kwargs)
+
+# wait for most 5s (arbitrary)
+MAX_TIMEOUT=2
+
+class SteppedBytesIO(io.BytesIO):
+    """
+    SteppedBytesIO provide a way to stop the closing of the IO N-1 times, closing on the Nth time.
+    """
+    def __init__(self, initial_bytes: bytes = b'', steps: int = 1):
+        self.steps = steps
+
+    def close(self):
+        if self.steps <= 0:
+            super(io.BytesIO, self).close()
+        else:
+            self.steps -= 1
+
+# TODO: create fake endpoint class (or similar) so the input can be stepped
+
+# https://github.com/python-lsp/python-lsp-server/blob/develop/pylsp/python_lsp.py#L58
+def server_output(
+        input: typing.BinaryIO,
+        timeout: float | None = MAX_TIMEOUT) -> typing.Tuple[typing.BinaryIO, MALLSPServer, TimeoutError | None]:
+    intermediary = SteppedBytesIO()
+    ls = MALLSPServer(input, intermediary)
+    time_out_err = None
+
+    async def run_server():
+        ls.start()
+
+    try:
+        server_future = run_server()
+        asyncio.run(asyncio.wait_for(server_future, 1))
+    except TimeoutError as e:
+        ls.m_exit()
+        time_out_err = e
+    except Exception as e:
+        intermediary.close()
+        raise e
+
+    return intermediary, ls, time_out_err
